+++ sdk/go/packet/doc.go
@@ -1,3 +1,3 @@
-// A Pulumi package for creating and managing X cloud resources.
+// A Pulumi package for creating and managing Packet cloud resources.
 //
 package packet